LayoutTests/fast/dom/MutationObserver/mutation-observer-constructor-expected.txt

Test the constructor of CyberKitMutationObserver

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.CyberKitMutationObserver is non-null.
PASS typeof CyberKitMutationObserver.prototype.observe is "function"
PASS typeof CyberKitMutationObserver.prototype.disconnect is "function"
PASS typeof observer.observe is "function"
PASS typeof observer.disconnect is "function"
PASS new MutationObserver({ handleEvent: function() {} }) threw exception TypeError: Argument 1 ('callback') to the MutationObserver constructor must be a function.
PASS new MutationObserver({}) threw exception TypeError: Argument 1 ('callback') to the MutationObserver constructor must be a function.
PASS new MutationObserver(42) threw exception TypeError: Argument 1 ('callback') to the MutationObserver constructor must be a function.
PASS new MutationObserver("foo") threw exception TypeError: Argument 1 ('callback') to the MutationObserver constructor must be a function.
PASS successfullyParsed is true

TEST COMPLETE
